[Silver III] Title: 주유소, Time: 440 ms, Memory: 36740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -5,18 +5,18 @@ public class Main{
     public static void main(String[] args) throws IOException {
     	BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     	int n = Integer.parseInt(br.readLine());
-    	int road[] = new int[n-1];
-    	int city[] = new int[n];
+    	long road[] = new long[n-1];
+    	long city[] = new long[n];
     	StringTokenizer st;
     	st = new StringTokenizer(br.readLine());
     	for(int i = 0; i < n-1; i++) // 0 1 2 
-    		road[i] = Integer.parseInt(st.nextToken());
+    		road[i] = Long.parseLong(st.nextToken());
     	st = new StringTokenizer(br.readLine());
     	for(int i = 0; i < n; i++) // 0 1 2 3
-    		city[i] = Integer.parseInt(st.nextToken());
+    		city[i] = Long.parseLong(st.nextToken());
     	
-    	int cost = 0;
-    	int min = city[0]; //5
+    	long cost = 0;
+    	long min = city[0]; //5
     	//가격에 초점 맞춰 비교 
     	//더 싸네? 여기서 도로 길이만큼 더 넣자
     	for(int i = 0; i < n-1; i++) {
